// frontend/src/app/core/api/model/api-response-course-session-response.ts
/**
 * Training school system Backend API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ErrorDetail } from './error-detail';
import { CourseSessionResponse } from './course-session-response';


export interface ApiResponseCourseSessionResponse { 
    success?: boolean;
    message?: string;
    errorCode?: string;
    data?: CourseSessionResponse;
    errors?: Array<ErrorDetail>;
    timestamp?: string;
}